feat: TransactionManager, ConnectionManager 분리
